refactor: move types into src tree
