+++ src/types/colors.d.ts
