+++ src/types/test/global.ts
